Update to udp2beh.py to be compatible with odin_II
odin_II doesn't allow the outputs to be overridden with conditional if statements.  Thus it produces the following error.

	"Note that Odin II does not currently support combinational a = ? overiding for if and case blocks."

I added code to change subsequent ifs in the outputAssignment block to be else ifs.  I was concerned that this would essentially reverse the order of precedence, but I believe this is fine if the original table was fully specified.  (Please let me know if you think I'm wrong about it.)